+++ charts/victoria-metrics-k8s-stack/CHANGELOG.md
@@ -278,7 +278,7 @@
 ![AppVersion: v1.102.1](https://img.shields.io/static/v1?label=AppVersion&message=v1.102.1&color=success&logo=)
 ![Helm: v3](https://img.shields.io/static/v1?label=Helm&message=v3&color=informational&logo=helm)
 
-- TODO
+- Fix templating of labels for `VMAlertmanager` CRD.
 
 ## 0.25.4