+++ charts/victoria-metrics-k8s-stack/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 - TODO
 
+## 0.28.1
+
+**Release date:** 2024-11-05
+
+![AppVersion: v1.106.0](https://img.shields.io/static/v1?label=AppVersion&message=v1.106.0&color=success&logo=)
+![Helm: v3](https://img.shields.io/static/v1?label=Helm&message=v3&color=informational&logo=helm)
+
+- TODO
+
 ## 0.28.0
 
 **Release date:** 2024-11-05